Move get_user_version, set_user_version to sqlite.db_upgrades.version
